requirements.txt
pika==1.3.2
python-dotenv==1.2.2
celery==5.5.3
